tests/integration/test_key_service.py
"""
Testes de integração para KeyService.
Requer DATABASE_URL configurado (SQLite em memória para testes).
"""
import pytest
import asyncio
from app.database.engine import init_db


@pytest.mark.asyncio
async def test_placeholder():
    # Substitua por testes reais com fixture de session
    assert True
