requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
aiomysql>=0.2.0
alembic>=1.13.0
pydantic-settings>=2.0.0
qrcode[pil]>=7.4.2
pillow>=10.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.1
